fix: (un)marshalling tuple and dictionary unwrapping for Typescript (#1717)
