+++ src/helpers/index.ts
@@ -7,3 +7,4 @@ export * from './Constraints';
 export * from './ConstrainHelpers';
 export * from './PresetHelpers';
 export * from './DependencyHelpers';
+export * from './FilterHelpers';
